+++ NEWS.md
@@ -3,6 +3,9 @@
 * cache and batch weather data acquisition by year (#1);
 * allow setting cache path using option `landisutils.cache.path`;
 * change arguments `start` and `end` in `prep_*_weather()` to be integer years;
+* use `R6` classes to track simulation input files;
+* implemented scenario replication (#3);
+* run landis in background process via `callr`;
 
 # landisutils 0.0.1